feat(wamn-iq82.1): a contract integer is int32, and a revision carries its width
An integer that an operation contract declares now maps to int32, so it is a
number in TypeScript and on the wire. A model column keeps its own type, and a
declared int64 stays the opaque string alias.

A revision may be int32 or int64. The contract field rule, the column rule, the
WIT request member, the update and delete codecs and the refusal detail all
follow the revision column. Introspection reads the literal default of an int4
column, which an int32 revision needs.

The three generated packages regenerate. The only change is the page limit,
which is now a number.

+++ crates/schema/generator/src/client_fields.rs
@@ -301,7 +301,10 @@ fn schema_field(schema: &Value, mut path: String, required: bool, hints: &[&Fiel
             result.type_name = hints.iter().find(|hint| hint.path == path).map_or_else(
                 || {
                     match ty {
-                        Some("integer") => "int64",
+                        // An integer the operation contract declares is int32.
+                        // A model column keeps its own type through its hint,
+                        // so a bigint column still reads as int64 here.
+                        Some("integer") => "int32",
                         Some("number") => "float64",
                         Some("boolean") => "boolean",
                         _ => match schema.get("format").and_then(Value::as_str) {
